requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
Automat==0.7.0
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.5
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==6.0.0
nbconvert==5.4.1
nbformat==4.4.0
nltk==3.4
notebook==5.7.4
numpy==1.16.1
pandas==0.24.1
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pkg-resources==0.0.0
pluggy==0.9.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyenchant==2.0.0
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.2.2
pymongo==3.7.2
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
pytest==4.3.0
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.0
qtconsole==4.4.3
queuelib==1.5.0
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
Scrapy==1.6.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.7.3
tabulate==0.8.3
terminado==0.8.1
testpath==0.4.2
tornado==4.5.3
traitlets==4.3.2
Twisted==18.9.0
typed-ast==1.3.1
urllib3==1.24.1
w3lib==1.20.0
watson-developer-cloud==2.8.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.48.0
widgetsnbextension==3.4.2
wordcloud==1.5.0
wrapt==1.11.1
zope.interface==4.6.0
